Add function to achieve unequal block size (16*16 and 32*32)
* Add global boolean variable to see whether the user choose to use the
unequal size feature.
* Add in GUI in the block size selection a option for user to choose the
unequal size.
* Add warning for user to not to choose the "less tile repetition"
function as well as the "unequal size" function
* Add function in Tiler.cpp to loop the tileIndex to see whether there
are 4 blocks locate as the square shape use the same tile. If yes, then
combine the 4 blocks into 1, and use a larger size of the same tile to
replace.
* Add print out to see the basic information of the target processing
(e.g. block size, the number of block, the number of tiles and the time
for processing).